\name{SamplesModelCatalytic}
\alias{SamplesModelCatalytic}
\alias{SamplesModelCatalyticYear}
\alias{SamplesModelConstPrev}
\alias{SamplesModelDensity}
\docType{data}
\title{
  MCMC Samples for the Four Models Compared in the Paper 
}
\description{
  MCMC Samples for the Four Models Compared in the Paper.
}
\usage{

data("SamplesModelCatalytic")

data("SamplesModelCatalyticYear")

data("SamplesModelConstPrev")

data("SamplesModelDensity")

}
\format{
  These datasets are objects of class \code{mcmc.list} (package
  \code{rjags}).
}
\details{

  These datasets have been generated by the function \code{coda.samples}
  for each model discussed in the paper. The JAGS code used to generate
  these samples is describe on the help page of
  \code{ModelCatalytic}.  We show how we obtained to the samples in the
  examples for the main model \code{ModelCatalyticYear}. Note that the
  results stored in these datasets do not correspond to those presented
  in the paper: they correspond to the results obtained with the
  simulated dataset \code{BruData} in this package (which we simulated
  since we did not have the authorization to include the actual data in
  the package, see the help page of \code{BruData}).
  
}
\references{

  Calenge C., Lambert S., Petit E., Thebault A., Gilot-Fromont E., Toigo
  C., Rossi S. (in prep.) Estimating disease prevalence and temporal
  dynamics using biased capture serological data in a wildlife
  reservoir: the example of brucellosis in Alpine ibex (Capra
  ibex). Submitted to Preventive Veterinary Medicine.

}
\examples{

## Load the jags code for the model
data(ModelCatalytic)

\dontrun{

library(rjags)

obmcy <- jags.model(textConnection(ModelCatalyticYear),
                    data=BruData, n.chains = 5,
                    inits =
                    list(e=as.numeric(BruData$BacterialNpositive>0),
           eta=0.1))

SamplesModelCatalyticYear <- coda.samples(obmcy, n.iter=5000, thin=5,
                                          variable.names = c("pitu","pitm","pit", "gammat",
                                                             "eta","beta0","beta1","beta2","beta3",
                                                             "sigmaf","probad","probai",
                                                             "wt","lambdat", "omegat"))

## This results in the same dataset as
## data(SamplesModelCatalyticYear)

}


}
\keyword{datasets}
